+++ lib/kb_gtdbtk/kb_gtdbtkServer.py
@@ -338,10 +338,10 @@ class Application(object):
         self.serverlog.set_log_level(6)
         self.rpc_service = JSONRPCServiceCustom()
         self.method_authentication = dict()
-        self.rpc_service.add(impl_kb_gtdbtk.run_kb_gtdbtk,
-                             name='kb_gtdbtk.run_kb_gtdbtk',
+        self.rpc_service.add(impl_kb_gtdbtk.run_kb_gtdbtk_classify_wf,
+                             name='kb_gtdbtk.run_kb_gtdbtk_classify_wf',
                              types=[dict])
-        self.method_authentication['kb_gtdbtk.run_kb_gtdbtk'] = 'required'  # noqa
+        self.method_authentication['kb_gtdbtk.run_kb_gtdbtk_classify_wf'] = 'required'  # noqa
         self.rpc_service.add(impl_kb_gtdbtk.status,
                              name='kb_gtdbtk.status',
                              types=[dict])